+++ src/main/java/ua/com/khrypko/family/budget/repository/FamilyRepository.java
@@ -6,5 +6,5 @@ import ua.com.khrypko.family.budget.entity.user.Family;
 /**
  * Created by maks on 19.11.17.
  */
-public interface FamilyRepository extends JpaRepository<Family, Integer> {
+public interface FamilyRepository extends JpaRepository<Family, Long> {
 }
