+++ src/main/java/ua/com/khrypko/family/budget/service/user/FamilyService.java
@@ -1,6 +1,6 @@
 package ua.com.khrypko.family.budget.service.user;
 
-import ua.com.khrypko.family.budget.dto.FamilyDTO;
+import ua.com.khrypko.family.budget.dto.user.FamilyDTO;
 import ua.com.khrypko.family.budget.entity.user.Family;
 
 /**
@@ -8,15 +8,15 @@ import ua.com.khrypko.family.budget.entity.user.Family;
  */
 public interface FamilyService {
 
-    Family getFamily(int id);
-    Family getFamilyByUser(int userId);
+    Family getFamily(long id);
+    Family getFamilyByUser(long userId);
 
-    FamilyDTO getFamilyDTO(int id);
-    FamilyDTO getFamilyDTOByUser(int userId);
+    FamilyDTO getFamilyDTO(long id);
+    FamilyDTO getFamilyDTOByUser(long userId);
 
     Family createFamily(FamilyDTO familyDTO);
-    Family addUser(int familyId, int userId);
-    Family removeUser(int familyId, int userId);
+    Family addUser(long familyId, long userId);
+    Family removeUser(long familyId, long userId);
     Family updateFamily(FamilyDTO familyDTO);
 
 }
